add tests for creating bookings
